+++ src/pages/unAuth/IsSignedInLanding.js
@@ -1,51 +1,95 @@
-import React, { useEffect } from 'react';
-import { SwishSpinner } from 'react-spinners-kit';
+import React from 'react';
 
-import { Gooddollar } from '../../components/pages/landing/Gooddollar';
-import { Panel } from '../../components/common/panel';
+import { FractalVerification } from '../../components/pages/landing/FractalVerification';
+import { Tabs } from '../../components/common/Tabs';
+import { wallet } from '../..';
 
-export const IsSignedInLanding = ({
-  hasApplied,
-  userData,
-  showGooddollarVerification,
-  setShowGooddollarVerification,
-}) => {
-  useEffect(() => {
-    if (window.location.href.includes('?login')) {
-      setShowGooddollarVerification(true);
-    }
-  }, [setShowGooddollarVerification, hasApplied]);
+export const IsSignedInLanding = ({ activeTabIndex, setActiveTabIndex }) => {
+  const TabsData = [
+    {
+      name: 'Connect Wallet',
+      header: (
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          fill="none"
+          viewBox="0 0 24 24"
+          stroke-width="1.5"
+          stroke="currentColor"
+          class="w-12 h-12 stroke-white"
+        >
+          <path
+            stroke-linecap="round"
+            stroke-linejoin="round"
+            d="M21 12a2.25 2.25 0 00-2.25-2.25H15a3 3 0 11-6 0H5.25A2.25 2.25 0 003 12m18 0v6a2.25 2.25 0 01-2.25 2.25H5.25A2.25 2.25 0 013 18v-6m18 0V9M3 12V9m18 0a2.25 2.25 0 00-2.25-2.25H5.25A2.25 2.25 0 003 9m18 0V6a2.25 2.25 0 00-2.25-2.25H5.25A2.25 2.25 0 003 6v3"
+          />
+        </svg>
+      ),
+      content: (
+        <div className="mt-8">
+          <p className="text-s mb-5">
+            Near wallet is required. Be sure you connect the wallet you want to
+            use for governance. Get a new wallet for free if you don't already
+            have one (or need an alt).
+          </p>
+          <button
+            onClick={() => wallet.signIn()}
+            type="button"
+            className="flex items-center mx-auto rounded border border-transparent bg-indigo-600 w-60 py-1.5 text-xs font-medium text-white shadow-sm hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2"
+          >
+            <p className="mx-auto w-[fit-content]">Connect</p>
+          </button>
+        </div>
+      ),
+    },
+    {
+      name: 'Face Scan',
+      header: (
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          fill="none"
+          viewBox="0 0 24 24"
+          stroke-width="1.5"
+          stroke="currentColor"
+          class="w-12 h-12 stroke-white"
+        >
+          <path
+            stroke-linecap="round"
+            stroke-linejoin="round"
+            d="M15.182 15.182a4.5 4.5 0 01-6.364 0M21 12a9 9 0 11-18 0 9 9 0 0118 0zM9.75 9.75c0 .414-.168.75-.375.75S9 10.164 9 9.75 9.168 9 9.375 9s.375.336.375.75zm-.375 0h.008v.015h-.008V9.75zm5.625 0c0 .414-.168.75-.375.75s-.375-.336-.375-.75.168-.75.375-.75.375.336.375.75zm-.375 0h.008v.015h-.008V9.75z"
+          />
+        </svg>
+      ),
+      content: <FractalVerification />,
+    },
+    {
+      name: 'Mint SBT',
+      header: (
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          fill="none"
+          viewBox="0 0 24 24"
+          stroke-width="1.5"
+          stroke="currentColor"
+          class="w-12 h-12 stroke-white"
+        >
+          <path
+            stroke-linecap="round"
+            stroke-linejoin="round"
+            d="M9 12.75L11.25 15 15 9.75M21 12c0 1.268-.63 2.39-1.593 3.068a3.745 3.745 0 01-1.043 3.296 3.745 3.745 0 01-3.296 1.043A3.745 3.745 0 0112 21c-1.268 0-2.39-.63-3.068-1.593a3.746 3.746 0 01-3.296-1.043 3.745 3.745 0 01-1.043-3.296A3.745 3.745 0 013 12c0-1.268.63-2.39 1.593-3.068a3.745 3.745 0 011.043-3.296 3.746 3.746 0 013.296-1.043A3.746 3.746 0 0112 3c1.268 0 2.39.63 3.068 1.593a3.746 3.746 0 013.296 1.043 3.746 3.746 0 011.043 3.296A3.745 3.745 0 0121 12z"
+          />
+        </svg>
+      ),
+      content: <FractalVerification />,
+    },
+  ];
 
   return (
     <>
-      {hasApplied === null && (
-        <div className="w-[fit-content] pt-20 space-x-3 mx-auto">
-          <div className="w-[fit-content] mx-auto">
-            <SwishSpinner color="blue" />
-          </div>
-          <p className="animate text-sm mt-2 text-gray-500">
-            Loading your account details{' '}
-            <span className="animate-pulse">...</span>
-          </p>
-        </div>
-      )}
-      <Panel
-        open={showGooddollarVerification}
-        onClose={() => setShowGooddollarVerification(false)}
-        title="Get Verified and Apply for Community SBT"
-      >
-        <div className="max-w-6xl mx-auto px-2">
-          <Gooddollar
-            setShowGooddollarVerification={setShowGooddollarVerification}
-          />
-        </div>
-      </Panel>
-      {hasApplied === true && (
-        <>
-          {/* {userData?.status === "Application Denied" && <p>TO DO</p>} */}
-          {/* {userData?.status === "Approved" && <CheckSbtTokenStatus />} */}
-        </>
-      )}
+      <Tabs
+        tabs={TabsData}
+        activeTabIndex={activeTabIndex}
+        setActiveTabIndex={() => null}
+      />
     </>
   );
 };
